+++ src/plone/restapi/services/actions/get.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+from Products.CMFCore.interfaces import IActionCategory
+from Products.CMFCore.utils import getToolByName
 from plone import api
 from plone.restapi.services import Service
 from zope.i18n import translate
@@ -6,12 +8,23 @@ from zope.i18n import translate
 
 class ActionsGet(Service):
 
+    @property
+    def all_categories(self):
+        portal_actions = getToolByName(self.context, 'portal_actions')
+        categories = []
+        for id, obj in portal_actions.objectItems():
+            if IActionCategory.providedBy(obj):
+                categories.append(id)
+        return categories
+
     def reply(self):
         context_state = api.content.get_view(name='plone_context_state',
                                              context=self.context,
                                              request=self.request)
+        categories = self.request.form.get('categories',
+                                           self.all_categories)
         data = {}
-        for category in ['object', 'object_buttons', 'user']:
+        for category in categories:
             category_action_data = []
             actions = context_state.actions(category=category)
             for action in actions: