+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 1.3.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add @actions endpoint to get portal_actions
+  [csenger]
 
 
 1.3.0 (2018-03-05)